requirements.txt
fastapi
uvicorn
websockets
deepgram-sdk
python-dotenv
openai
supabase
python-multipart